+++ client/src/components/Navbar/Navbar.module.css
@@ -1,0 +1,159 @@
+.logo {
+  width: 40px;
+}
+
+.icon {
+  width: 25px;
+  height: 25px;
+}
+
+.nav {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 10px 20px;
+  background-color: #031525;
+  color: #0f295c;
+  border-bottom: solid 0.5px #30363d;
+}
+
+.navBtns {
+  justify-content: space-between;
+  align-items: center;
+}
+
+.column {
+  z-index: 1000;
+  position: absolute;
+  top: 4.5rem;
+  right: 1rem;
+  width: 125px;
+  padding: 15px;
+  border-radius: 15px;
+  border: 1px solid lightgrey;
+  background-color: #031525;
+}
+.column::before {
+  content: "";
+  position: absolute;
+  top: -0.7rem;
+  right: 1.1rem;
+  width: 20px;
+  height: 20px;
+  transform: rotate(45deg);
+  background-color: #031525;
+  border-left: 1px solid lightgrey;
+  border-top: 1px solid lightgrey;
+}
+.li {
+  list-style: none;
+}
+
+.dropDownDiv {
+  display: flex;
+  color: #afb8c1;
+  margin-bottom: 10px;
+  padding-bottom: 5px;
+  border-bottom: solid 0.5px #30363d;
+  cursor: pointer;
+}
+
+.dropDownSpan {
+  margin-left: 3px;
+}
+
+.divLogout {
+  cursor: pointer;
+}
+
+.userPhoto {
+  cursor: pointer;
+  display: inline;
+  width: 49px;
+  border-radius: 10px;
+  border: solid 0.5px #30363d;
+  align-items: center;
+  margin-bottom: -14px;
+}
+
+.userPhoto:hover {
+  transition: 0.5s;
+  background-color: #1f2147;
+  border-color: white;
+  box-shadow: 0px 0px 10px #ffffff;
+}
+
+.userPhoto2 {
+  width: 25px;
+  border-radius: 10px;
+  border: solid 0.5px #30363d;
+}
+
+.connectBtn,
+.signOut,
+.premium,
+.chatBtn,
+.adminBtn,
+.groupBtn,
+.column {
+  background: transparent;
+  border: solid 0.5px #30363d;
+  border-radius: 10px;
+  margin-right: 10px;
+}
+.column:not(:hover),
+.connectBtn:not(:hover),
+.signOut:not(:hover),
+.premium:not(:hover),
+.chatBtn:not(:hover),
+.adminBtn:not(:hover),
+.groupBtn:not(:hover) {
+  transition: 0.5s;
+}
+.adminBtn:hover {
+  transition: 0.5s;
+  background-color: #1f2147;
+  border-color: red;
+  box-shadow: 0px 0px 10px rgb(195, 193, 193);
+}
+.adminBtn:hover {
+  transition: 0.5s;
+  background-color: #1f2147;
+  border-color: red;
+  box-shadow: 0px 0px 10px rgb(255, 0, 0);
+}
+
+.chatBtn:hover {
+  transition: 0.5s;
+  background-color: #1f2147;
+  border-color: rgb(21, 255, 0);
+  box-shadow: 0px 0px 10px #49f028;
+}
+
+.groupBtn:hover {
+  transition: 0.5s;
+  background-color: #1f2147;
+  border-color: rgb(132, 0, 255);
+  box-shadow: 0px 0px 10px #b128f0;
+}
+
+.connectBtn:hover {
+  transition: 0.5s;
+  background-color: #1f2147;
+  border-color: #00bfc7;
+  box-shadow: 0px 0px 10px #ffffff;
+}
+
+.premium:hover {
+  transition: 0.5s;
+  background-color: #1f2147;
+  border-color: yellow;
+  box-shadow: 0px 0px 10px #f0e628;
+}
+
+.signOut:hover {
+  transition: 0.5s;
+  background-color: #1f2147;
+  border-color: grey;
+  box-shadow: 0px 0px 10px #8f8f8f;
+}
